requirements.txt
virny
numpy==1.23.5
matplotlib~=3.6.2
pandas~=1.5.2
altair~=4.2.0
scikit-learn~=1.2.0
tqdm~=4.64.1
sklearn-utils
seaborn~=0.12.1
folktables~=0.0.11
python-dotenv~=1.0.0
pytest~=7.2.2
pymongo~=4.3.3
fairlearn~=0.9.0
aif360~=0.5.0